Treated dist-clean and better compilation

+++ buildthis.sh
@@ -40,7 +40,7 @@ while :; do
       echo 'ERROR: "--clean" requires a non-empty option argument.\n' >&2
       exit 1
       ;;
-    --distclean)
+    --distclean|--dist-clean)
       if [ ${2#--} != $2 ]; then
         distclean=1
         clean=1
@@ -53,13 +53,13 @@ while :; do
         continue
       fi
       ;;
-    --distclean=?*)
+    --distclean=?*|--dist-clean=?*)
       distclean=${1#*=} # Delete everything up to "=" and assign the remainder.
       if test "$distclean" -eq "1"; then
         clean=1
       fi
       ;;
-    --distclean=)   # Handle the case of an empty --distclean=
+    --distclean=|--dist-clean=)   # Handle the case of an empty --distclean=
       echo 'ERROR: "--distclean" requires a non-empty option argument.\n' >&2
       exit 1
       ;;
@@ -95,24 +95,20 @@ done
 # Set RootCore environment
 source ./setrootcore.sh --silent --no-env-setup
 
-# If we are using LCG numpy, then add it to environment now:
-source "$ROOTCOREBIN/../RootCoreMacros/retrieve_python_info.sh" --numpy-info
-test "x$NUMPY_LCG" != "x" && add_to_env PYTHONPATH "$PYTHON_NUMPY_PATH"
+# Compile
+test $clean -eq "1" && "$ROOTCOREBIN/bin/$ROOTCORECONFIG/rc" clean
+if test $distclean -eq "1"; then
+  echo "cleaning everything..."
+  rm -rf "$DEP_AREA" "$INSTALL_AREA"
+  # Remove old environment files (to be sure that we won't have old files on the environment):
+  for file in `find -L "$ROOTCOREBIN/.." -maxdepth 3 -mindepth 3 -path "*/cmt/*" -name "$BASE_NEW_ENV_FILE" `
+  do
+    test -x "$file" && rm "$file"
+  done
+fi
 
 # Now add the new environment files
 source ./setrootcore.sh --silent
-
-# Remove old environment files (to be sure that we won't have old files on the environment):
-NEW_ENV_FILE="$(basename "$NEW_ENV_FILE")"
-# Remove all environment files
-for file in `find -L "$ROOTCOREBIN/.." -maxdepth 3 -mindepth 3 -path "*/cmt/*" -name "$NEW_ENV_FILE" `
-do
-  test -x "$file" && rm "$file"
-done
-
-# Compile
-test $clean -eq "1" && "$ROOTCOREBIN/bin/$ROOTCORECONFIG/rc" clean
-test $distclean -eq "1" -a -n "$ROOTCOREBIN" && rm -r "$DEP_AREA" "$INSTALL_AREA"
 
 if ! "$ROOTCOREBIN/bin/$ROOTCORECONFIG/rc" compile
 then
